completed create project route. implemented sample frontend index for testing. resolved fstring usage for improved security. updated getresearchers route to properly display all projects in user project list. implemented audio file storage system. implemented user check for dockerRebuild

+++ dockerRebuild.sh
@@ -1,9 +1,15 @@
 #! /bin/bash
 
-# Stop and remove all containers, networks, and volumes
-echo "Removing Docker containers, networks, and volumes..."
-docker-compose down -v
+echo "Doing this will remove all Docker Containers, Networks, Volumes and Saved Data"
+read -p "Are you Sure? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    # Stop and remove all containers, networks, and volumes
+    echo "Removing Docker containers, networks, and volumes..."
+    docker-compose down -v
 
-# Rebuild and start the Docker containers
-echo "Rebuilding and starting Docker containers..."
-docker-compose up --build
+    # Rebuild and start the Docker containers
+    echo "Rebuilding and starting Docker containers..."
+    docker-compose up --build
+fi
